+++ docs/changelog.md
@@ -16,6 +16,10 @@ Each revision is versioned by the date of the revision.
 
 - Added relation endpoints reference page.
 
+## 2026-05-27
+
+- Added optional `dns-record` relation to publish DNS A records for HAProxy hostnames.
+
 ## 2026-05-25
 
 - docs: Updated the home page to implement the new pattern for the "In this documentation" section.